requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0

# Segmentation
segmentation-models-pytorch>=0.3.3

# OCR
transformers>=4.35.0
sentencepiece>=0.1.99

# Geospatial
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
rasterio>=1.3.0

# Evaluation
icdar21-mapseg-eval

# Utilities
matplotlib>=3.7.0
tqdm>=4.65.0
pandas>=2.0.0
scikit-image>=0.21.0
scipy>=1.11.0
